move address-autocomplete to own file
